make xmlenc optional dependency, and refactor to split encryption into EncryptedAssertionBuilder
