+++ spec/lib/saml_idp/assertion_builder_spec.rb
@@ -12,13 +12,6 @@ module SamlIdp
       Saml::XML::Namespaces::AuthnContext::ClassRef::PASSWORD
     }
     let(:expiry) { 3*60*60 }
-    let (:encryption_opts) do
-      {
-        cert: Default::X509_CERTIFICATE,
-        block_encryption: 'aes256-cbc',
-        key_transport: 'rsa-oaep-mgf1p',
-      }
-    end
     subject { described_class.new(
       reference_id,
       issuer_uri,
@@ -35,29 +28,6 @@ module SamlIdp
       Timecop.travel(Time.zone.local(2010, 6, 1, 13, 0, 0)) do
         subject.raw.should == "<Assertion xmlns=\"urn:oasis:names:tc:SAML:2.0:assertion\" ID=\"_abc\" IssueInstant=\"2010-06-01T13:00:00Z\" Version=\"2.0\"><Issuer>http://sportngin.com</Issuer><Subject><NameID Format=\"urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress\">foo@example.com</NameID><SubjectConfirmation Method=\"urn:oasis:names:tc:SAML:2.0:cm:bearer\"><SubjectConfirmationData InResponseTo=\"123\" NotOnOrAfter=\"2010-06-01T13:03:00Z\" Recipient=\"http://saml.acs.url\"></SubjectConfirmationData></SubjectConfirmation></Subject><Conditions NotBefore=\"2010-06-01T12:59:55Z\" NotOnOrAfter=\"2010-06-01T16:00:00Z\"><AudienceRestriction><Audience>http://example.com</Audience></AudienceRestriction></Conditions><AttributeStatement><Attribute Name=\"email-address\" NameFormat=\"urn:oasis:names:tc:SAML:2.0:attrname-format:uri\" FriendlyName=\"emailAddress\"><AttributeValue>foo@example.com</AttributeValue></Attribute></AttributeStatement><AuthnStatement AuthnInstant=\"2010-06-01T13:00:00Z\" SessionIndex=\"_abc\"><AuthnContext><AuthnContextClassRef>urn:oasis:names:tc:SAML:2.0:ac:classes:Password</AuthnContextClassRef></AuthnContext></AuthnStatement></Assertion>"
       end
-    end
-
-    it "builds encrypted XML" do
-      builder = described_class.new(
-        reference_id,
-        issuer_uri,
-        name_id,
-        audience_uri,
-        saml_request_id,
-        saml_acs_url,
-        algorithm,
-        authn_context_classref,
-        expiry,
-        encryption_opts
-      )
-      encrypted_xml = builder.encrypt
-      encrypted_xml.should_not match(audience_uri)
-
-      encrypted_doc = Nokogiri::XML::Document.parse(encrypted_xml)
-      encrypted_data = Xmlenc::EncryptedData.new(encrypted_doc.at_xpath('//xenc:EncryptedData', Xmlenc::NAMESPACES))
-      decrypted_assertion = encrypted_data.decrypt(builder.encryption_key)
-      decrypted_assertion.should == subject.raw
-      decrypted_assertion.should match(audience_uri)
     end
   end
 end
